1873-04-22 Tests for Magazine Guns
The Board met on April 22nd, 1873. 
Present: All the members and the recorder.

The Board resolved to test the Ward-Burton magazine-arm, which happened on April 23rd. Unstated is that they would also test Lt. Metcalfe's detachable magazine for the Springfield, and Dr. Elliot's magazine. The tests decided upon were not captured in the minutes, but the May 5th report recaps the tests, which were necessarily different for each gun owing to their different capabilities.