1873-04-23 Tests for Facility of Manipulation
The Board met on April 23rd, 1873.
Present: All the members and the recorder.

The Board resolved to try the arms for facility of manipulation by persons not experts. The details of the tests were not captured in the minutes but were explained in the May 5th report. These tests were conducted between April 24th and May 3rd.