1873-05-05 Clarification of test for observing escaping gas
The final report of May 5th, 1873 offers clarification of how the paper was used in Test I - Defective Cartridges, compared to the test detailed in the meeting minutes. The statement prescribing the use of paper has been moved to clarify that this step applied to each element of Test I, not simply Test I.3.